Wifi: Clear scan caches while loading configured networks
Framework stores the scan caches for each configured network in a hash
map with network id as key and never clear them. If removing a network
and then re-enabling wifi, framework will re-assigned network id for
each configured network while loading them, however, the scan caches
still keeps the obsolete network id as its key.

This change is to clear the scan caches when framework loads configured
networks, which makes sure the network id in the scan caches is up to
date.

CRs-Fixed: 999659
Change-Id: I1a6cdb2871e6f58f7aa5d233517d5983be91a7df